requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.13.1
fonttools==4.49.0
formulaic==1.0.1
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
idna==3.6
interface-meta==1.3.0
ipykernel==6.25.2
ipython==8.16.1
ISLP==0.3.22
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
l0bnb==1.0.0
lifelines==0.28.0
lightning-utilities==0.10.1
llvmlite==0.42.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.59.0
numpy==1.24.2
opencv-python==4.10.0.82
packaging==23.2
pandas==1.5.3
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
progressbar2==4.3.2
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyclustering==0.10.1.2
pygam==0.9.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-utils==3.8.2
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.0
torchinfo==1.8.0
torchmetrics==1.3.1
torchsummary==1.5.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.11.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.8
wrapt==1.16.0
yarl==1.9.4
yellowbrick==1.5
